The toolbar needs refresh when a mutation occurs on one of its parents.
Consider such a case: the editor is hidden (for example, the display
property is set to none), and then the window is resized by the user.
The toolbar is refreshed on resize event. However, since the editor
is hidden, the recalculated width of the toolbar is 0. Therefore, when
the editor is displayed again, the toolbar needs to be refreshed again
to get the correct width.