Fix the exception thrown from InputManager.prototype._onFocus.
In the situation that the textarea is not autofocus and there's an image
in the textarea, when the page is refreshed and the textarea is still
not focus, immediately click the image in the textarea, an exception
will be thrown from the following statement in
InputManager.prototype._onFocus():

range = _this.editor.selection._selection.getRangeAt(0);

because there's nothing actually selected yet in the editor at the moment.
The click event to select the image is triggered after the focus event.